MSim: always enable activity disabler
Because MSim setting is a persist variable, and a phone can change
from non-MSim to MSim by flashing a ROM with added MSim support,
we should always run the disabler to make sure the state is always
correct.

The disabler has been extended to also enable the related
component which switching from non-MSim to MSim or vice-versa.

The rename is to avoid the activity not being called due to it
being disabled by itself earlier.

Change-Id: Ibe976bd01ac23cc6665900dafdc158d945480474